Fix Experiment Manager Decimal/float type conversion errors
- Convert Decimal database values to float for mathematical operations
- Fix total_return and daily_return calculations in portfolio.py
- Fix shares and unrealized_percent calculations in main.py
- Ensure all financial calculations work with PostgreSQL Decimal types

+++ experiment_manager/main.py
@@ -300,7 +300,7 @@ class ExperimentManagerService:
             # Рассчитываем текущую P&L
             current_value = position['shares'] * current_price
             unrealized_pnl = current_value - position['position_size']
-            unrealized_percent = (unrealized_pnl / position['position_size']) * 100
+            unrealized_percent = float(unrealized_pnl / position['position_size']) * 100
 
             logger.debug(f"Position update: {position['ticker']}")
             logger.debug(f"  Current price: ${current_price:.2f}")
