+++ experiment_manager/portfolio.py
@@ -145,7 +145,7 @@ class PortfolioManager:
                 'realized_pnl_today': snapshot['realized_pnl_today'],
                 'realized_pnl_total': snapshot['realized_pnl_total'],
                 'daily_return': float(((current_value / snapshot['total_value']) - 1) * 100) if snapshot['total_value'] > 0 else 0,
-                'total_return': ((current_value / self.config.INITIAL_CAPITAL) - 1) * 100,
+                'total_return': float(((current_value / self.config.INITIAL_CAPITAL) - 1) * 100),
                 'available_cash': snapshot['cash_balance'],
                 'last_updated': snapshot['timestamp']
             }
@@ -216,7 +216,7 @@ class PortfolioManager:
                 return False, f"Would violate cash reserve: ${cash_after_position} < ${min_cash_reserve}"
 
             # Проверка 6: Daily loss limit
-            daily_loss_percent = abs(portfolio['realized_pnl_today']) / portfolio['total_value'] * 100
+            daily_loss_percent = float(abs(portfolio['realized_pnl_today']) / portfolio['total_value']) * 100
             if daily_loss_percent >= self.config.DAILY_LOSS_LIMIT_PERCENT:
                 return False, f"Daily loss limit reached: {daily_loss_percent:.1f}% >= {self.config.DAILY_LOSS_LIMIT_PERCENT}%"
 
@@ -233,7 +233,7 @@ class PortfolioManager:
             sp500_price = self.market_data.get_benchmark_price('SPY')
 
             # Рассчитываем параметры позиции
-            shares = execution_data['position_size'] / execution_data['execution_price']
+            shares = float(execution_data['position_size']) / float(execution_data['execution_price'])
             commission = self.config.calculate_commission(execution_data['position_size'])
             total_cost = execution_data['position_size'] + commission
 
@@ -329,7 +329,7 @@ class PortfolioManager:
             sp500_return = 0
             alpha = 0
             if experiment['sp500_entry'] and sp500_exit:
-                sp500_return = ((sp500_exit / experiment['sp500_entry']) - 1) * 100
+                sp500_return = float((sp500_exit / experiment['sp500_entry']) - 1) * 100
                 alpha = return_percent - sp500_return
 
             # Время удержания
@@ -487,7 +487,7 @@ class PortfolioManager:
         """Проверяет превышение дневного лимита потерь"""
         try:
             portfolio = self.get_portfolio_status()
-            daily_loss_percent = abs(portfolio['realized_pnl_today']) / portfolio['total_value'] * 100
+            daily_loss_percent = float(abs(portfolio['realized_pnl_today']) / portfolio['total_value']) * 100
 
             return daily_loss_percent >= self.config.DAILY_LOSS_LIMIT_PERCENT
